corrección de numero de ejercicio
